implemented sort page before recipe list and tile functionality
